修复: 编译警告 - 格式字符串类型不匹配
- 修复camera_data_write.c中的格式字符串警告
- 将size_t类型强制转换为uint16_t以匹配%04X格式
- 确保编译无警告，同时保持功能正常

测试验证:
- 编译无警告
- 写入和读取程序功能正常
- 数据完整性验证通过

+++ test/src/camera_data_write.c
@@ -337,9 +337,9 @@ int main(int argc, char* argv[]) {
         
         printf("\n=== 写入完成 ===\n");
         printf("总共写入文件数: %d\n", file_count);
-        printf("EEPROM使用地址范围: 0x%04X - 0x%04X\n", 
-               EEPROM_START_ADDRESS, EEPROM_START_ADDRESS + sizeof(index_header_t) + 
-               (file_count * sizeof(file_index_t)) + files[file_count - 1].size - 1);
+        printf("EEPROM使用地址范围: 0x%04X - 0x%04X\n",
+               EEPROM_START_ADDRESS, (uint16_t)(EEPROM_START_ADDRESS + sizeof(index_header_t) +
+               (file_count * sizeof(file_index_t)) + files[file_count - 1].size - 1));
     }
     
     // 清理资源
